Move GriddedPSFModel tests to new module
