+++ photutils/psf/tests/test_models.py
@@ -3,17 +3,12 @@
 Tests for the models module.
 """
 
-from itertools import product
-
 import numpy as np
 import pytest
 from astropy.modeling.models import Gaussian2D, Moffat2D
-from astropy.nddata import NDData
 from numpy.testing import assert_allclose
 
-from photutils.psf import (FittableImageModel, GriddedPSFModel,
-                           IntegratedGaussianPRF, PRFAdapter)
-from photutils.segmentation import SourceCatalog, detect_sources
+from photutils.psf import FittableImageModel, IntegratedGaussianPRF, PRFAdapter
 from photutils.utils._optional_deps import HAS_SCIPY
 
 
@@ -113,207 +108,6 @@ class TestFittableImageModel:
                              np.ones((2, 2, 2)), 2.1, np.nan, (1, np.inf)]:
             with pytest.raises(ValueError):
                 FittableImageModel(data, oversampling=oversampling)
-
-
-@pytest.fixture(name='psfmodel')
-def fixture_griddedpsf_data():
-    psfs = []
-    y, x = np.mgrid[0:101, 0:101]
-    for i in range(16):
-        theta = i * 10.0 * np.pi / 180.0
-        g = Gaussian2D(1, 50, 50, 10, 5, theta=theta)
-        m = g(x, y)
-        psfs.append(m)
-
-    xgrid = [0, 40, 160, 200]
-    ygrid = [0, 60, 140, 200]
-    grid_xypos = list(product(xgrid, ygrid))
-
-    meta = {}
-    meta['grid_xypos'] = grid_xypos
-    meta['oversampling'] = 4
-
-    nddata = NDData(psfs, meta=meta)
-    psfmodel = GriddedPSFModel(nddata)
-
-    return psfmodel
-
-
-class TestGriddedPSFModel:
-    """
-    Tests for GriddPSFModel.
-    """
-
-    def test_gridded_psf_model(self, psfmodel):
-        keys = ['grid_xypos', 'oversampling']
-        for key in keys:
-            assert key in psfmodel.meta
-        assert len(psfmodel.meta) == 2
-        assert len(psfmodel.meta['grid_xypos']) == 16
-        assert psfmodel.oversampling == 4
-        assert psfmodel.meta['oversampling'] == psfmodel.oversampling
-        assert psfmodel.data.shape == (16, 101, 101)
-
-    @pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-    def test_gridded_psf_model_basic_eval(self, psfmodel):
-        y, x = np.mgrid[0:100, 0:100]
-        psf = psfmodel.evaluate(x=x, y=y, flux=100, x_0=40, y_0=60)
-        assert psf.shape == (100, 100)
-
-    @pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-    def test_gridded_psf_model_eval_outside_grid(self, psfmodel):
-        y, x = np.mgrid[-50:50, -50:50]
-        psf1 = psfmodel.evaluate(x=x, y=y, flux=100, x_0=0, y_0=0)
-        y, x = np.mgrid[-60:40, -60:40]
-        psf2 = psfmodel.evaluate(x=x, y=y, flux=100, x_0=-10, y_0=-10)
-        assert_allclose(psf1, psf2)
-
-        y, x = np.mgrid[150:250, 150:250]
-        psf3 = psfmodel.evaluate(x=x, y=y, flux=100, x_0=200, y_0=200)
-        y, x = np.mgrid[170:270, 170:270]
-        psf4 = psfmodel.evaluate(x=x, y=y, flux=100, x_0=220, y_0=220)
-        assert_allclose(psf3, psf4)
-
-    @pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-    def test_gridded_psf_model_interp(self, psfmodel):
-        # test xyref length
-        with pytest.raises(TypeError):
-            psfmodel._bilinear_interp([1, 1], 1, 1, 1)
-
-        # test if refxy points form a rectangle
-        with pytest.raises(ValueError):
-            xyref = [[0, 0], [0, 1], [1, 0], [2, 2]]
-            zref = np.ones((4, 4, 4))
-            psfmodel._bilinear_interp(xyref, zref, 1, 1)
-
-        # test if xi and yi are outside of xyref
-        xyref = [[0, 0], [0, 1], [1, 0], [1, 1]]
-        zref = np.ones((4, 4, 4))
-        with pytest.raises(ValueError):
-            psfmodel._bilinear_interp(xyref, zref, 100, 1)
-        with pytest.raises(ValueError):
-            psfmodel._bilinear_interp(xyref, zref, 1, 100)
-
-        # test non-scalar xi and yi
-        idx = [0, 1, 4, 5]
-        xyref = np.array(psfmodel.grid_xypos)[idx]
-        psfs = psfmodel.data[idx, :, :]
-        val1 = psfmodel._bilinear_interp(xyref, psfs, 10, 20)
-        val2 = psfmodel._bilinear_interp(xyref, psfs, [10], [20])
-        assert_allclose(val1, val2)
-
-    def test_gridded_psf_model_invalid_inputs(self):
-        data = np.ones((4, 3, 3))
-
-        # check if NDData
-        with pytest.raises(TypeError):
-            GriddedPSFModel(data)
-
-        # check PSF data dimension
-        with pytest.raises(ValueError):
-            GriddedPSFModel(NDData(np.ones((3, 3))))
-
-        # check that grid_xypos is in meta
-        meta = {'oversampling': 4}
-        nddata = NDData(data, meta=meta)
-        with pytest.raises(ValueError):
-            GriddedPSFModel(nddata)
-
-        # check grid_xypos length
-        meta = {'grid_xypos': [[0, 0], [1, 0], [1, 0]],
-                'oversampling': 4}
-        nddata = NDData(data, meta=meta)
-        with pytest.raises(ValueError):
-            GriddedPSFModel(nddata)
-
-        # check if grid_xypos is a regular grid
-        meta = {'grid_xypos': [[0, 0], [1, 0], [1, 0], [3, 4]],
-                'oversampling': 4}
-        nddata = NDData(data, meta=meta)
-        with pytest.raises(ValueError):
-            GriddedPSFModel(nddata)
-
-        # check that oversampling is in meta
-        meta = {'grid_xypos': [[0, 0], [0, 1], [1, 0], [1, 1]]}
-        nddata = NDData(data, meta=meta)
-        with pytest.raises(ValueError):
-            GriddedPSFModel(nddata)
-
-        # check oversampling is a scalar
-        meta = {'grid_xypos': [[0, 0], [0, 1], [1, 0], [1, 1]],
-                'oversampling': [4, 4]}
-        nddata = NDData(data, meta=meta)
-        with pytest.raises(ValueError):
-            GriddedPSFModel(nddata)
-
-    @pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-    def test_gridded_psf_model_eval(self, psfmodel):
-        """
-        Create a simulated image using GriddedPSFModel and test
-        the properties of the generated sources.
-        """
-        shape = (200, 200)
-        data = np.zeros(shape)
-        eval_xshape = (np.ceil(psfmodel.data.shape[2]
-                               / psfmodel.oversampling)).astype(int)
-        eval_yshape = (np.ceil(psfmodel.data.shape[1]
-                               / psfmodel.oversampling)).astype(int)
-
-        xx = [40, 50, 160, 160]
-        yy = [60, 150, 50, 140]
-        zz = [100, 100, 100, 100]
-        for xxi, yyi, zzi in zip(xx, yy, zz):
-            x0 = np.floor(xxi - (eval_xshape - 1) / 2.0).astype(int)
-            y0 = np.floor(yyi - (eval_yshape - 1) / 2.0).astype(int)
-            x1 = x0 + eval_xshape
-            y1 = y0 + eval_yshape
-
-            x0 = max(x0, 0)
-            y0 = max(y0, 0)
-            x1 = min(x1, shape[1])
-            y1 = min(y1, shape[0])
-
-            y, x = np.mgrid[y0:y1, x0:x1]
-            data[y, x] += psfmodel.evaluate(x=x, y=y, flux=zzi, x_0=xxi,
-                                            y_0=yyi)
-
-        segm = detect_sources(data, 0.0, 5)
-        cat = SourceCatalog(data, segm)
-        orients = cat.orientation.value
-        assert_allclose(orients[1], 50.0, rtol=1.0e-5)
-        assert_allclose(orients[2], -80.0, rtol=1.0e-5)
-        assert 88.3 < orients[0] < 88.4
-        assert 64.0 < orients[3] < 64.2
-
-    def test_copy(self, psfmodel):
-        flux = psfmodel.flux.value
-        new_model = psfmodel.copy()
-        new_model.flux = 100
-        assert new_model.flux.value != flux
-        assert new_model._data_input is psfmodel._data_input
-
-    def test_deepcopy(self, psfmodel):
-        flux = psfmodel.flux.value
-        new_model = psfmodel.deepcopy()
-        new_model.flux = 100
-        assert new_model.flux.value != flux
-
-    @pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-    def test_cache(self, psfmodel):
-        for x, y in psfmodel.grid_xypos:
-            psfmodel.x_0 = x
-            psfmodel.y_0 = y
-            psfmodel(0, 0)
-            psfmodel(1, 1)
-
-        assert psfmodel._cache_info().hits == 16
-        assert psfmodel._cache_info().misses == 16
-        assert psfmodel._cache_info().currsize == 16
-
-        psfmodel.clear_cache()
-        assert psfmodel._cache_info().hits == 0
-        assert psfmodel._cache_info().misses == 0
-        assert psfmodel._cache_info().currsize == 0
 
 
 @pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
